fix(sentry): tracing module import (#248)
