+++ .env
@@ -1,0 +1,3 @@
+NEXT_PUBLIC_SENTRY_DSN='https://b3d1f140b36f441da22791ca98210c95@o1029417.ingest.sentry.io/6052500'
+NEXT_PUBLIC_GOOGLE_ANALYTICS='G-KLS272T65L'
+NEXT_PUBLIC_HASH_VERSION=$HASH_VERSION
